Adding ClusterTest for Mssql (for optimistic locking strategy)
