+++ .github/workflows/ci.yml
@@ -18,6 +18,6 @@ jobs:
         cache: 'maven'
 
     - name: Run all tests
-      run: mvn -B -Pcompatibility clean test --file pom.xml
+      run: mvn -B -Pall-tests clean test --file pom.xml
       env:
         TZ: UTC
