+++ db-scheduler/src/test/java/com/github/kagkarlsson/scheduler/Utils.java
@@ -1,0 +1,12 @@
+package com.github.kagkarlsson.scheduler;
+
+public class Utils {
+
+  public static void sleep(int millis) {
+    try {
+      Thread.sleep(millis);
+    } catch (InterruptedException e) {
+      throw new RuntimeException(e);
+    }
+  }
+}
